requirements.txt
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
fastjsonschema==2.14.5
Flask==1.1.2
Flask-API==2.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.5
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
